+++ Tests/Data/Types/Basic/IntegerTypeTest.php
@@ -9,13 +9,11 @@ class IntegerTypeTest extends TestCase
 {
     public function testGetRequirementsOptions()
     {
-        $integerType = new IntegerType();
         $this->assertTrue(is_array(IntegerType::getRequirementsOptions()));
     }
 
     public function testCheckPassInteger()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $actual         = IntegerType::check($expected);
 
@@ -25,7 +23,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassNumerical()
     {
-        $integerType    = new IntegerType();
         $expected       = "5";
         $actual         = IntegerType::check($expected);
         $expected       = (int) $expected;
@@ -35,7 +32,7 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataTypeException
      * @expectedExceptionMessage value must be an integer
      */
     public function testCheckIntegerFail()
@@ -47,7 +44,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassIntegerWithMinRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $requirements   = array('min' => 2);
         $actual         = IntegerType::check($expected, $requirements);
@@ -57,12 +53,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage value must be greater than or equal to
      */
     public function testCheckFailIntegerWithMinRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $requirements   = array('min' => 4);
         $actual         = IntegerType::check($expected, $requirements);
@@ -70,7 +65,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassIntegerWithMinAndMaxRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $requirements   = array('min' => 2, 'max' => 3);
         $actual         = IntegerType::check($expected, $requirements);
@@ -80,12 +74,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage and greater than or equal
      */
     public function testCheckFailIntegerWithmaxAndMinRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 7;
         $requirements   = array('min' => 3, 'max' => 5);
         $actual         = IntegerType::check($expected, $requirements);
@@ -95,12 +88,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage and less than or equal to
      */
     public function testCheckFailIntegerWithMinAndMaxRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 2;
         $requirements   = array('min' => 3, 'max' => 5);
         $actual         = IntegerType::check($expected, $requirements);
@@ -118,12 +110,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage value must be less than or equal to
      */
     public function testCheckFailIntegerWithMaxRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 10;
         $requirements   = array('max' => 7);
         $actual         = IntegerType::check($expected, $requirements);
@@ -131,7 +122,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassIntegerWithValuesRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $requirements   = array('values' => array(1,3,5,7,9));
         $actual         = IntegerType::check($expected, $requirements);
@@ -141,12 +131,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage value must be one of
      */
     public function testCheckFailIntegerWithValuesRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 2;
         $requirements   = array('values' => array(1,3,5,7,9));
         $actual         = IntegerType::check($expected, $requirements);
@@ -154,7 +143,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassIntegerWithOddRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
         $requirements   = array('odd' => true);
         $actual         = IntegerType::check($expected, $requirements);
@@ -164,12 +152,11 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage value must be an odd number
      */
     public function testCheckFailIntegerWithOddRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 4;
         $requirements   = array('odd' => true);
         $actual         = IntegerType::check($expected, $requirements);
@@ -177,7 +164,6 @@ class IntegerTypeTest extends TestCase
 
     public function testCheckPassIntegerWithEvenRequirements()
     {
-        $integerType    = new IntegerType();
         $expected       = 4;
         $requirements   = array('even' => true);
         $actual         = IntegerType::check($expected, $requirements);
@@ -187,7 +173,7 @@ class IntegerTypeTest extends TestCase
     }
 
     /**
-     * @expectedException Ucc\Exception\InvalidDataTypeException
+     * @expectedException Ucc\Exception\Data\InvalidDataValueException
      * @expectedExceptionMessage value must be an even number
      */
     public function testCheckFailIntegerWithEvenRequirements()
@@ -200,7 +186,6 @@ class IntegerTypeTest extends TestCase
 
     public function testIsPass()
     {
-        $integerType    = new IntegerType();
         $expected       = 3;
 
         $this->assertTrue(IntegerType::is($expected));
@@ -208,7 +193,6 @@ class IntegerTypeTest extends TestCase
 
     public function testIsFail()
     {
-        $integerType    = new IntegerType();
         $expected       = 'abc';
 
         $this->assertFalse(IntegerType::is($expected));
