+++ Ucc/Data/Types/Basic/StringType.php
@@ -3,7 +3,9 @@
 namespace Ucc\Data\Types\Basic;
 
 use Ucc\Data\Types\TypeInterface;
-use Ucc\Exception\InvalidDataTypeException;
+use Ucc\Exception\Data\InvalidDataTypeException;
+use Ucc\Exception\Data\InvalidDataValueException;
+use Ucc\Exception\Data\InvalidDataException;
 
 /**
  * Ucc\Data\Types\Basic\StringType
@@ -37,7 +39,7 @@ class StringType implements TypeInterface
      * @param   mixed   $value          Value to be checked
      * @param   array   $requirements   Additional constraints
      * @return  mixed   Cleared value
-     * @throws  InvalidDataTypeException
+     * @throws  InvalidDataTypeException | InvalidDataValueException
      */
     public static function check($value, array $requirements = array())
     {
@@ -53,7 +55,7 @@ class StringType implements TypeInterface
                 $error = 'value must be one of: '
                     . implode(', ', $requirements['values']);
 
-                throw new InvalidDataTypeException($error);
+                throw new InvalidDataValueException($error);
             }
         // Check min and max length if both options are supplied
         } elseif (isset($requirements['min']) && isset($requirements['max'])) {
@@ -62,7 +64,7 @@ class StringType implements TypeInterface
                     . $requirements['min'] . ' to '
                     . $requirements['max'] . ')';
 
-                throw new InvalidDataTypeException($error);
+                throw new InvalidDataValueException($error);
             }
         }
         // Check min
@@ -71,7 +73,7 @@ class StringType implements TypeInterface
                 $error = 'value length must be greater than or'
                     . ' equal to ' . $requirements['min'];
 
-                throw new InvalidDataTypeException($error);
+                throw new InvalidDataValueException($error);
             }
         }
         // Check max
@@ -80,7 +82,7 @@ class StringType implements TypeInterface
                 $error = 'value length must be less than or'
                     . ' equal to ' . $requirements['max'];
 
-                throw new InvalidDataTypeException($error);
+                throw new InvalidDataValueException($error);
             }
         }
 
@@ -95,13 +97,12 @@ class StringType implements TypeInterface
      * @param   array   $requirements   Additional constraints
      * @return  boolean                 True if value is of a given type and
      *                                  meets requirements
-     * @throws  InvalidDataTypeException
      */
     public static function is($value, array $requirements = array())
     {
         try {
             self::check($value, $requirements);
-        } catch (InvalidDataTypeException $e) {
+        } catch (InvalidDataException $e) {
             return false;
         }
