+++ Ucc/Exception/Data/InvalidDataTypeException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Ucc\Exception\Data;
+
+use Ucc\Exception\Data\InvalidDataException;
+
+/**
+ * This exception is thrown when a data type does not match expected type.
+ *
+ * @author Kris Rybak <kris@krisrybak.com>
+ */
+class InvalidDataTypeException extends InvalidDataException
+{
+}
