+++ Ucc/Exception/Data/InvalidDataValueException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Ucc\Exception\Data;
+
+use Ucc\Exception\Data\InvalidDataException;
+
+/**
+ * This exception is thrown when a data value does not match requirements specified.
+ *
+ * @author Kris Rybak <kris@krisrybak.com>
+ */
+class InvalidDataValueException extends InvalidDataException
+{
+}
